requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.5
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
sqlparse==0.4.3
swagger-spec-validator==3.0.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.10.0
